requirements.txt
beautifulsoup4
requests
openai
python-dotenv
tiktoken
langchain
nltk
pinecone-client